requirements.txt
pydantic-ai
tavily-python
python-dotenv